Rename maven-setup references to setup-maven
