--- maven-setup/action.yml
@@ -1,82 +1,0 @@
-name: Maven Setup Action
-description: Setup action to capture Maven Build Scans metadata
-author: Gradle team
-
-inputs:
-  develocity-access-key:
-    description: 'Develocity access key. Should be set to a secret containing the Develocity Access key'
-    required: false
-  develocity-token-expiry:
-    description: 'Develocity short-lived access tokens expiry in hours (Default 2 hours)'
-    required: false
-  capture-strategy:
-    description: 'Build Scan capture strategy (ALWAYS, ON_FAILURE, ON_DEMAND)'
-    required: false
-    default: ALWAYS
-  capture-unpublished-build-scans:
-    description: 'Whether to enable or not unpublished Build Scans capture'
-    required: false
-    default: true
-  capture-build-scan-links:
-    description: 'Whether to enable or not Build Scan links capture'
-    required: false
-    default: true
-  job-name:
-    description: 'Job name (used in summary and as artifact suffix)'
-    required: false
-  job-matrix:
-    description: 'Job matrix; this should not be overridden by users (INTERNAL).'
-    required: false
-    default: ${{ toJSON(matrix) }}
-  add-pr-comment:
-    description: 'Whether to comment or not the pull-request with Build Scan links'
-    default: 'true'
-  add-job-summary:
-    description: 'Whether to add or not a job summary to the GitHub workflow'
-    default: 'true'
-  add-project-id-in-job-summary:
-    description: 'Whether to add or not the project id in the job summary / comment'
-    default: 'false'
-  github-token:
-    description: 'The token used for Github API requests'
-    default: ${{ github.token }}
-    required: false
-  develocity-injection-enabled:
-    description: 'Whether to enable or not Develocity injection'
-    required: false
-  develocity-url:
-    description: 'Develocity server URL'
-    required: false
-  develocity-maven-extension-version:
-    description: 'Develocity Maven Extension version'
-    required: false
-  develocity-ccud-maven-extension-version:
-    description: 'Common Custom User Data Maven Extension version'
-    required: false
-  develocity-allow-untrusted-server:
-    description: 'Whether to allow communicating with untrusted server'
-    required: false
-  develocity-enforce-url:
-    description: 'Whether to enforce the URL over any defined locally to the project'
-    required: false
-  develocity-capture-file-fingerprints:
-    description: 'Whether to enable/disable capturing of file fingerprints. This setting is enabled by default.'
-    required: false
-    default: 'true'
-  develocity-custom-develocity-maven-extension-coordinates:
-    description: 'Will not inject the Develocity Maven extension if an extension with provided coordinates is found (value is an artifactId of the extension).'
-    required: false
-  develocity-custom-ccud-extension-coordinates:
-    description: 'Will not inject the CCUD extension if an extension with provided coordinates is found (value is an artifactId of the extension).'
-    required: false
-
-
-outputs:
-  build-metadata-file-path:
-    description: 'Path to the summary file build-metadata.json'
-
-runs:
-  using: node20
-  pre: ../dist/maven-setup/pre/index.js
-  main: ../dist/maven-setup/main/index.js
-  post: ../dist/maven-setup/post/index.js
