+++ package.json
@@ -23,9 +23,8 @@
     "format:test": "prettier --write **/__tests__/*.ts **/__tests__/**/*.ts",
     "lint": "eslint */src/**/*.ts",
     "package-maven-capture-extension": "mvn clean package -f maven-build-scan-capture-extension/pom.xml",
-    "package-maven-setup-pre": "ncc build maven-setup/src/pre.ts --out dist/maven-setup/pre",
-    "package-maven-setup-main": "ncc build maven-setup/src/main.ts --out dist/maven-setup/main",
-    "package-maven-setup-post": "ncc build maven-setup/src/post.ts --out dist/maven-setup/post",
+    "package-maven-setup-main": "ncc build setup-maven/src/main.ts --out dist/setup-maven/main",
+    "package-maven-setup-post": "ncc build setup-maven/src/post.ts --out dist/setup-maven/post",
     "package-maven-publish": "ncc build maven-publish-build-scan/src/main.ts --out dist/maven-publish-build-scan",
     "check": "npm-run-all --parallel format:write format:test lint",
     "package": "npm-run-all --parallel package-maven-*",
