+++ setup-maven/action.yml
@@ -77,5 +77,5 @@ outputs:
 
 runs:
   using: node20
-  main: ../dist/maven-setup/main/index.js
-  post: ../dist/maven-setup/post/index.js
+  main: ../dist/setup-maven/main/index.js
+  post: ../dist/setup-maven/post/index.js
